Remove host running check for system dump
The check for host running was added for the
operations on the dumps stored in the host
but in the case of request for offload The system dump
is ready to offload as soon as it is ready in the host.
The state of the host in the BMC may not be updated by
then, so it is not reliable to depend on the host state
in the BMC and if the host is not ready there will be an
error from PLDM.

This commit is for removing the host running check before
sending the initiate offload request to host. The check
for other operations like delete or initiate will not
be impacted

Signed-off-by: Dhruvaraj Subhashchandran <dhruvaraj@in.ibm.com>
Change-Id: I665485c7605b54fe4f361ec73e2bff9bba87c49e