Fix failure to create dump zip file as info.yaml file not found
Due to some reason info.yaml file is not found, and due to
that dump generation is halted.

Dump generation should not be halted if info.yaml file is
not generated nor found instead should log an error and
continue with zipping the dump file.

Tested:
info.yaml file found adding to dump

Signed-off-by: Marri Devender Rao <devenrao@in.ibm.com>
Change-Id: Ie7de5702469044cecfaf9db61967fe8b228c1ea0