Fix duplicate InterfacesRemoved signal
If the interface object is created with defer_emit, only the
InterfacesAdded signal is deferred. But it still emits InterfacesRemoved
signal. So when we inherit the entry class from multiple interfaces,
one of the interface can have the default or defer_emit and the other
interfaces should not emit any signal.

Tested and found only one InterfacesRemoved signal emitted on bmc dump
deletion. And tested the other dump creation and deletion and it works
as expected.

Signed-off-by: Deepa Karthikeyan <deepakala.karthikeyan@ibm.com>