Handle exceptions thrown during checkAndInitialize (#120)
If dump manager crashes core file will be generated but core dump will
not be created as the dump service that creates core dump is taking a
reset. We check for core files at the start of the dump service and
capture them in core dump.If there was an exception while capturing the
dump it was unhandled

Fix:
Added exception handling for the above scenario.

Signed-off-by: Deepa Karthikeyan <deepakala.karthikeyan@ibm.com>