dreport: Use Custom Package instead of Header Extension
HEADER_EXTENSION has only utility functions to add
headers for packaging the dump. It is more appropriate
to call the packaging script instead which would
handle the sub-logic of HEADER_EXTENSION within the broader
task of packaging the dumps.

IBM and OpenPower plugins are being moved to
ibm-openbmc/openpower-debug-collector repository.
So these scripts are being removed from here.

Tested: Verified that BMC Dump is getting created
using the modified dreport.

Change-Id: I6e4dd98f7474e1a47f25b1fed322553feb09a3eb
Signed-off-by: Gopichand Paturi <gopichandpaturi@gmail.com>